tradingview/requirements.txt
flask
gunicorn
python-binance
python-dotenv
redis